DAB: the OFDM transform through PFFFT, and a receiver that stops allocating 1.5 MB a frame
Profiled ON THE PHONE this time — simpleperf on the XCover decoding 7D (the app is
profileable since this morning), not the Mac's numbers transplanted:

    Fft::forward (radix-2, std::complex)   19.6 %   <- the single largest symbol
    DabReceiver::push                       15.1 %
    Resample24to2048::process               14.2 %
    Viterbi::decode (already NEON)          12.7 %

★★★ THE FFT IS NOW PFFFT where the build already links it — app, server, dab-offline
and test-dab-receiver define VIBE_DAB_PFFFT — and the 40-line radix-2 stays as the
default and the REFERENCE, so every vibe_dab_*.h still compiles with a two-line
g++ and links nothing (the header-only rule this file was written under). Same
convention both ways: e^{-2πi}, natural order, unscaled. After: the transform is
1.4 % of the profile (pffft_transform_internal), vibe-dab 55 % → 46.5 % of a core.
dab-offline decodes the capture to the identical 452 clean frames on either path;
all 20 server tests pass.

★★ push() allocated its frame buffer (~1.5 MB), the carrier arrays, the spectrum
and the soft bits afresh every frame — ten mallocs, page-faults and frees a
second on a phone. They are members now; assign() reuses the capacity.

★ derotate() walks every sample of every frame in double; the renormalisation
  every 1024 samples already bounds a float oscillator's drift, so the precision
  was bought and then thrown away. Float now.

★ 2 048 000 is offered in the RTL-SDR rate list. Stuart noticed it never was
  "for some reason" — and it is the one rate that lets the 2.4-vs-2.048 question
  be MEASURED on the XCover rather than argued: whether the 94 % delivery at
  2.048 was the dongle or the phone starving its USB reader. Offered, not
  defaulted; DAB still captures at 2.4.

+++ android/app/src/main/cpp/vibe_dab_ofdm.h
@@ -72,17 +72,21 @@ inline float offsetHz(float fractional, const Mode& m) { return fractional * flo
  *    difference between real time and not on a Pi. Renormalised every 1024 steps because the
  *    recurrence drifts in magnitude, which would otherwise scale the constellation slowly. */
 inline void derotate(Cplx* x, size_t n, double cyclesPerSample) {
+    /* ★ Float, not double: this walks every sample of every frame (2 M a second), and the
+     *  renormalisation every 1024 samples already bounds the drift a float oscillator has —
+     *  the double version bought precision the renorm then threw away. The step itself is
+     *  computed in double and rounded once. */
     const double w = -2.0 * M_PI * cyclesPerSample;
-    const double cs = std::cos(w), sn = std::sin(w);
-    double cr = 1.0, ci = 0.0;
+    const float cs = float(std::cos(w)), sn = float(std::sin(w));
+    float cr = 1.0f, ci = 0.0f;
     for (size_t i = 0; i < n; ++i) {
-        const double xr = x[i].re, xi = x[i].im;
-        x[i].re = float(xr * cr - xi * ci);
-        x[i].im = float(xr * ci + xi * cr);
-        const double nr = cr * cs - ci * sn, ni = cr * sn + ci * cs;
+        const float xr = x[i].re, xi = x[i].im;
+        x[i].re = xr * cr - xi * ci;
+        x[i].im = xr * ci + xi * cr;
+        const float nr = cr * cs - ci * sn, ni = cr * sn + ci * cs;
         cr = nr; ci = ni;
         if ((i & 1023u) == 1023u) {          // renormalise
-            const double mag = std::sqrt(cr * cr + ci * ci);
+            const float mag = std::sqrt(cr * cr + ci * ci);
             if (mag > 0) { cr /= mag; ci /= mag; }
         }
     }
